+++ engine-runner-bin/src/main.rs
@@ -144,9 +144,7 @@ mod tests {
 			"--btc.backup_rpc.basic_auth_user=user".to_string(),
 			"--btc.backup_rpc.basic_auth_password=password".to_string(),
 			// Sol
-			"--sol.rpc.ws_endpoint=ws://localhost:8899".to_string(),
 			"--sol.rpc.http_endpoint=http://localhost:8899".to_string(),
-			"--sol.backup_rpc.ws_endpoint=ws://localhost:8899".to_string(),
 			"--sol.backup_rpc.http_endpoint=http://localhost:8899".to_string(),
 			// p2p
 			format!("--p2p.node_key_file={some_file}"),